Feat: auth validations
