+++ src/routes/index.routes.js
@@ -1,10 +1,10 @@
 import express from 'express';
-import { adminProtect, protect } from '@/middleware/auth';
 import authRouter from './auth.routes';
 import dashboardRouter from './dashboard.routes';
 import questionRouter from './question.routes';
 import submissionRouter from './submission.routes';
 import userRouter from './user.routes';
+import { adminProtect, protect } from '@/middleware/auth';
 
 const router = express.Router();
