+++ src/services/auth.js
@@ -1,9 +1,9 @@
-import bcrypt from 'bcrypt';
 import crypto from 'crypto';
+import bcrypt from 'bcrypt';
 import createError from 'http-errors';
+import { sendMail } from './email';
 import { createUser, findOneAndUpdateUser, getOneUser } from '@/repository/user';
 import { decodeJwtToken } from '@/utils';
-import { sendMail } from './email';
 
 export const authRegister = async ({ name, email, password, university, members }) => {
   const encryptedPassword = await new Promise((resolve, reject) => {
